xtr_vobsub: allow track without CodecPrivate to be extracted
There's hardly a reason that it should not be allowed. Even if file
with such track is considered a bad one, allowing the extraction
would actually give a chance the user to fix it, let alone that
such track/file works fine on many players anyway. (In the other
words, CodecPrivate is somewhat optional for a VobSub track, and it
might even be completely ignored.)