merge: fix reversed attachment selection
If selection is reversed, `item_selector_<>::get()` returns the
default value for IDs that aren't part of the list (and should
therefore not be excluded). The default value for the attachment mode
is "don't attach".

Therefore the function checking whether or not an attachment ID has
been selected must take reversal into account.

Fixes #3427.