kax analyzer: handle "existing seek head too small" situation properly
Often the programs have to relocate the Master elements in which the
modifications were done. In that case the Seek Head elements must also
be updated to reflect to the Master elements' new positions.

The analyzer tries to re-use existing Seek Head elements at the start
of the file. If it is too small, it'll be relocated to the end of the
file & a new compact Seek Head will be created at the front
referencing the relocated one at the back. The newly created Seek Head
will be placed in the position of the relocated Seek Head.

Unfortunately it's possible that the space vacated by the relocated
Seek Head is so small that a newly created Seek Head doesn't fit into
it. In that case the analyzer would end up in an endless loop, trying
to write data to the end of the file until the drive's full or the
program's killed.

The solution is not to use that space (by overwriting it with an EBML
Void element) & trying to find a different place at the start. This
can be an existing EBML Void element big enough for it, or by
relocating another level 1 element (such as Info, Tracks, Chapters
etc.) & using the space freed up by that relocation.

Fixes #3338.