propedit: only require tracks to be present when modifying tracks
mkvpropedit always tries to find & read tracks as the chapter & tag
modification classes are derived from the track modification
class. However, when modifying chapters or tags the tracks are not
actually modified. Therefore mkvpropedit should only abort with an
error if tracks are not found & modification of track properties is
requested.

Fixes #3498.