VobSub reader: only probe if file extension is `.idx` or `.sub`
Otherwise other file types might be mis-detected as VobSub if a VobSub
with the same base name exists in the same directory. For example, if
`en.idx` & `en.sub` exist in a directory together with `en.ac3`, then
trying to identify `en.ac3` will try to detect it as VobSub, which
very well may succeed.

Fixes #3508.