GUI: use Fusion style on Windows & remove "disable dark mode" option
Qt 6.5 has been released, and with it proper support for Windows 10's
& 11's dark application mode, but only for those styles that support
it. "Fusion" is one such style, but the old "Windows Vista" that
MKVToolNix had been using until now isn't.

As this is proper support for the dark application mode, the option
for disabling said mode has been removed — also due to it not being
easy to really disable the mode in Qt, nor to retrieve the default
light palette if Windows' dark application mode is on.

Fixes #3264 and others.