GUI: Windows: restore custom palette-based dark mode for Windows 10
Qt 6.5.0 does not support Windows' dark mode on Windows 10
itself. Therefore my custom palette-based variant must be used there.

On Windows 11 it is left to Qt.

The option for disabling the dark mode remains removed as it isn't
possible to disable it in Qt on Windows 11 anymore.

Fixes #3517. See also #3264.