GUI: cache language dialog after first use
Creating a new instance takes a noticeable amount of time for setting
up the all the combo boxes which have more than 10.000 entries in
total. This commit changes the behavior so that the dialog is only
created once: the first time the user wants to edit a
language. Afterwards the instance is cached & reused.

Changes to the preferences that alter the settings concerning the
language dialog will cause the cached instance to be discarded.