GUI: language dialog: re-initialize component controls on opening dialog properly
When following these steps:

1. select multiple tracks
2. open language dialog
3. select anything in the drop-down boxes or enter something in the
   line edits
4. hit cancel
5. open the language dialog again

the values selected/entered in step 3 were still selected, even if the
multiple selected tracks had different languages. Additionally the
state of the dialog was always that the value is invalid.

This wasn't intentional; the controls should all be reset to having
nothing selected/entered.

This could also happen under certain circumstances with single
selections.

Part of the fix of #3532.