GUI: Windows: always use custom light & dark palettes & let user chose which to use
This unifies color handling on Windows. First of all, it uses slightly
modified palettes directly from Qt 6.5.0, keeping the fix for the
barely legible dark blue links in dark mode. The light palette is
unmodified.

Second, this adds an option in the preferences, letting the user chose
which palette to apply: either the light one, the dark one, or one
based on system settings.

The system settings now only take this flag into account, nothing
else (e.g. not high contrast settings).

The logic which palette is used is rather simple:

• If the user has selected either light or dark in MKVToolNix GUI's
  settings, use that one.

• Otherwise look at the application color mode registry key & derive
  which palette to use from it.

Fixes #3528.