GUI: don't always set last open dir in saveTextToFile()
The last open dir is something only an action by the user should
change. The function always set the last open dir, even though it is
also used for temporary files during identification and in other
circumstances when the directory wasn't chosen by the user, though.

The result was that the when adding a file the last open dir was set
to the dir the file comes from. Immediately afterwards the file
identification created a temporary file, though, which caused the last
open dir to be set to the temporary directory again.