NEWS: add info about the Boost.Filesystem bug wrt. SMB1 & creating directories
Before trying to open/create the destination file `mkvmerge` always
tries to ensure that the destination directory exists by calling the
function `boost::filesystem::create_directories()`. Unfortunately the
Boost.Filesystem library contained [a
bug](https://github.com/boostorg/filesystem/issues/282) in recent
versions that manifested when determining directory entry information
on a network share connected via SMB1.

The result was that Boost.Filesystem thought that an existing
directory either didn't exist or that the filesystem entry is not
actually a directory. Boost.Filesystem would then try to create the
directory, which failed as it existed already, causing
`create_directories()` to fail which in turn caused `mkvmerge` to
abort.

This bug was fixed in Boost.Filesystem's version 1.82.0, which
MKVToolNix for Windows will be built with from now on.

Fixes #3547.