GUI: merge: optionally ensure at least one track of each type is enabled
The GUI will now check the state of the 'track enabled' flag of all
video, audio & subtitle tracks when starting to multiplex or adding a
job to the job queue. If there's no track with its flag turned on for
a type, the GUI will turn on the flag for the first track of the
type.

This behavior can be turned off in the preferences.

Implements #3535.