GUI: don't try showing Qt dialogs from non-main threads
Happened in a helper function for writing text to a file. The error is
now passed upwards via an exception. At a proper place the exception
is caught & passed on via Qt's signals/slots mechanism to the main
thread, from which displaying a dialog is safe.

Part of the fix of #3558 & #3561.