merge: fix sorting tracks by type if tracks are deactivated
If some tracks are deactivated a reader's `m_reader_packetizers`
member will only contain the packetizers for the tracks that are
actually copied. In that case the index of the packetizer into the
vector is not the track's ID wrt. the track order.

The actual ID is stored in the packetizer's track info structure as
`m_ti.m_id`.

The result of the bug was that, when certain tracks were disabled,
track indexes used for the track order were wrong, causing the actual
track order to be wrong, too.

Part of the fix of #3567.