merge: fix assigning track numbers in track order when auto-ordering
Before the fix the track numbers of tracks in the destination file
were assigned as soon as the packetizer was created based on the
requested order of tracks (e.g. via the `--track-order` option). The
automatic ordering of tracks by their types uses the same mechanism as
`--track-order`.

Unfortunately the decisions for the track order required that all
packetizers had already been created. This resulted in a chicken & egg
problem: track order data generation relied on packetizers having been
created which in turn tried to use track order data during their
creation.

Now track numbers are simply assigned in the order the packetizers are
created (which already happened in the right order).

Part of the fix of #3567.