GUI: mux: fix deriving "forced display" flag from file name
In order to keep the number of command line arguments low, the GUI
only adds the ones for setting the "forced display" flag if its new
value is different than its previous one. Therefore when deriving
"forced display" = "yes" from the file name the GUI must not also set
the internal "that flag was set" data. Otherwise it'll think the old &
new states are identical, leading it to not adding the appropriate
command line flags.

See #3586.