AV1: Create Dolby Vision configuration record from ITU-T T.35 metadata OBUs
Currently for raw OBU bitstream input only.
The metadata is present in metadata OBUs with metadata_type = 4.

The format of the AV1 RPUs contains the size bytes, and consists of the original
RPU with shifted bytes, so it must be shifted back to the original payload first.
The method was found in existing vendor Linux kernel codebases.