DTS reader: skip leading EXSS if following core found
The sampling frequency in cores & in EXSS sub-stream elements may be
different. If a core is present, its sampling frequency must be used
for timestamp calculation; otherwise timestamps as well as the track's
overall duration will be wrong.

Therefore look for a core within the first probing buffer. If one is
found, start processing from there instead of from the first header
found.

Fixes #3602