merge: don't apply pixel aspect ratio on top of `--aspect-ratio-factor` if given
Due to how track order handling works, the function for setting the
display dimensions of a track is called twice if a track order is
present. That means that the function should be idempotent & not
change anything the second time around.

Unfortunately this wasn't the case if an aspect ratio factor was given
for a video track. In that case the result of the calculation is
stored back in a variable that is used in said calculation, meaning
the inputs for the calculation were different the second time
around. This led to the pixel aspect ratio being applied on top of the
given aspect ratio of the command line.

For example, with a pixel resolution of 720x576 (pixel aspect ratio of
5/4 or 1.25) and a given aspect ratio factor of 1/1, nothing should
change at all; the display dimensions should equal the pixel
dimensions at 720x576.

Instead, the calculation went something like this:

1. 720 (pixel width) * 1/1 (aspect ratio factor given on the command
   line) / 576 = 5/4; store that value in the same variable that
   previously held the aspect ratio factor given on the command line

2. 720 (pixel width) * 5/4 (the value from step 1) / 576 = 900/576

As stated initially this only happened when calling the function more
than once, which in turn only happened if a track order was
given. Unfortunately version 77 introduced sorting the tracks in the
output file by their type if no track order was given by the user,
which implicitly creates said track order, effectively triggering this
issue every time an aspect ratio factor was given.

That being said, even before version 77 it also always happened each
time `mkvmerge` was run from MKVToolNix GUI as the GUI always adds an
explicit track order.