MPEG TS reader: Blu-rays: skip TP_extra_headers properly
See Blu-ray specs section 6.2.1.

On Blu-rays the MPEG transport streams consist of a sequence of
so-called "source packets". A "source packet" consists of a four-byte
long `TP_extra_header` structure followed by the "transport packet"
that we're actually interested in. The latter is the one that starts
with a `0x47` byte.

The MPEG TS reader contains code to find the start of the "transport
packet" by looking for that `0x47` byte & then reading 192-byte long
chunks starting from there. This works well as long as the
`TP_extra_header` structure itself doesn't contain a `0x47`. As that
structure contains some type of timestamp, this is very much possible.

If this is the case, the detection of the start of the "transport
packet" is thrown off. Often a resync is attempted, during which at
a handful of "transport packets" are skipped, leading to a loss of
frames in the content (often video, which means dropping everything
before the second key frame).

Fixes #3632.