HEVC packetizer: only flush parser object once
When appending the HEVC packetizers of all the tracks in the appended
chain use the same HEVC ES parser object. When the one file is
finished, the packetizer is flushed, which in turn flushes the ES
parser. This is fine the first time the ES parser is flushed.

Now the core realizes that the first file is empty & starts reading
data from the appended file in order to determine the lowest
timestamp. The appended packetizer will add data to the same ES parser
object.

The core still reads from the first source file, though, especially if
there are more packetizers for the file (e.g. audio packetizers). That
file will run into "empty" more often for the original video
packetizer, causing more call to that packetizer's flush function.

The first file's packetizer will then flush the ES parser, which
already contains data from the second file. Those frames are then
queued in the first file's packetizer. However, the core already
considers the first file's packetizer to be done & won't pull from it
again.

The result is that data from the second file ends up in the first
file's packetizer's queue that's never processed further. It's still
there when mkvmerge finishes.

The fix is to only call the ES parser's flush function once per call
to a packetizer's flush function.

Fixes #3479.