MPEG TS: only create PCM converter if muxing of track is requested
Since v81 packets are processed if a converter exists, too, even if no
packetizer exists. This is required for combining Dolby vision layers
from different tracks as there's only be one track for which a
packetizer exists while the same converter instance is used for both
of them.

In turn this means that a converter should either be ready to deal
with not having a packetizer or it should not be created in the first
place. Not creating it saves a couple of CPU cycles as there's no
function call to the converter for each frame anymore.

This also adjusts the teletext subtitle conversion converter which
also relied on the packetizer having been created. Unfortunately this
converter is created so early during PMT parsing that not creating it
isn't really an option.

Fixes #3645.