mkvmerge: add hack writing `BlockAddID` elements even if set to default value
The `BlockAddID` sub-element of the `BlockMore` is a mandatory
element (in EBML speak: has a `minOccur` of 1) with a default value of
1. This means that the element not being present within a
`BlockMore` element is semantically identical to `BlockAddID` being
present & set to 1. In both cases a player must use the value 1.

There are players that don't support this properly. For those this
hack can be turned on, causing mkvmerge to write `BlockAddID` whenever
it writes a `BlockMore` element, even if the `BlockAddID` element's
value is 1.

This is a workaround for player issues mentioned in #3643.