fix compatibility with C++20
In C++20 `std::string` is `constexpr`. Therefore `fmt::format()` tries
to check the format at compile time, requiring its first argument to
be known at compile time as well. If it isn't (e.g. when it's the
result of a function call such as `gettext()`), it must be marked as
to be checked at runtime instead.