avoid warning about comparing different `enums` being deprecated
Due to historical reasons the `OSType` that gets returned from
`QOperatingSystemVersion::current().type()` is not the same one as
`QOperatingSystemVersion::OSType` but a base class's
type (`QOperatingSystemVersionBase::OSType`). The Qt developers keep
both instances in sync to ensure you can compare them.

However, in C++20 comparing two `enums` of different types is
deprecated, and warnings are emitted. The Qt developers seem to be
aware of this & plan to remove one of the two types in Qt 7 but cannot
do so during 6.x as it would break at least ABI & possibly API, too.

A well-places static cast gets rid of the warning & is still safe due
to the developers keeping the values of both `enums` identical.