merge: use immediate timestamp application mode for subtitle tracks
Otherwise frames use full queueing, which doesn't work well with
tracks for which frames are sparse — as is often the case for
subtitles. The result was that frames were queued until the next one
arrived. If the next one arrived very late, the formerly queued one
would only be written at that point in time & therefore might end up
mixed in with frames from other tracks with much higher timestamps.

partially fixes #3711