configure: fix setting `HAVE_QTDBUS` if module is found
The m4 macro `AC_DEFINE()` seems to write the definition into a file
that's later evaluated and from which `config.h` is built. This
temporary file must reside in the proper directory, otherwise
`configure` will only know that `HAVE_QTDBUS` might be defined but not
that it should actually be set.

The function checking for Qt6 uses a different temporary directory for
its checks whether `qmake` can compile with certain modules. For the
invocations it changes the current working directory to the temporary
directory. If `AC_DEFINE()` is used while the working directory is
changed, this will break the mechanism described above.

The solution is to only use `AC_DEFINE()` once the current working
directory's been changed back to the original one.

fixes #3744