More explicit control over use of QtDbus
Relates to #2886

I'm packaging mkvtoolnix for an environment that doesn't use dbus
and I found myself wanting more explicit control of whether or not
dbus is enabled.

This patch adds the following conifugre options:
    --enable-dbus=auto (default)
    --enable-dbus / --enable-dbus=yes
    --disable-dbus / --enable-dbus=no

The default auto mode matches the existing behavior, including dbus
support if it's found on the machine, and proceeding without it
otherwise.

If it's explicitly enabled, not finding dbus results in a fatal build
configuration error.

If it's explicitly disabled, it can be built without dbus support even
on machines where dbus is available.