Matroska reader: hard-limit buffered memory to 128 MB
When appending files, the core calls the reader with `force == true`
when looking for the smallest timestamp of each track in the appended
file. If there's a track with very few packets in total this means
that the Matroska reader would buffer all the data from the start of
the appended file until the first packet of each track has been found.

The most pathological case is one of forced-only subtitles, which
might contain as few as six entries in total over the length of a four
hour movie in the case of "Lord of the Rings — Return of the
King (extended cut)".

Fixes #3771.