AAC: LOAS/LATM: fix implicit audioMuxVersionA value
If the audioMuxVersion field within StreamMuxConfig() is set to 0, the
audioMuxVersionA field is not present but its value is implicitely set
to 0.

So far this path wasn't covered and the value of m_audio_mux_version_a
was undefined in such a case. In worst case this leads to a failed LATM
header recognition. In the particular case this led to a false positive
ADTS recognition as a consequence.

Hence set m_audio_mux_version_a to 0, if the audioMuxVersionA field is
not present.

Fixes an additional aspect of #3727.