HEVC/H.265 parser: fix picture order count calculation
This diff looks bigger than it actually is. Mainly it moves saving the
current frame's picture order count as the previous picture order
count from the "non-IDR only" case to always do it. Effectively what
happened was that it wasn't stored in the case of IDR_W_RADL &
IDR_N_NLP slices, which combined with certain following slices might
break the calculation for following slices.

fixes #3775