kax analyzer: fix replacing element at end followed by only void elements
When the element to replace is located at (or rather, near) the end of
the file & is only followed by one or more void elements, the code
must not assume that there's a non-void located afterwards, which it
did, causing the crash when looking for that non-existing element.

The fix is to apply the same logic as when there's no other element
following at all: just truncate the file where the element to replace
is located & return.

fixes #3802