PCM packetizer: always create 40ms long frames
The old code attempted to keep the size of frames identical to what
they were in the source files, as long as they were long enough (more
than 4ms). This would mean that pathetic framing in the source, such
as several seconds long frames, would be kept as-is.

Part of the fix of #6149