MP4 reader: fix exorbitant memory usage for PCM audio tracks
In MP4/QuickTime files PCM audio data is stored the following way:

- a lot of samples in big chunks of consecutive data
- however, only one entry in the "sample duration" table (`stts` atom)
  with a duration of 1 sample but a huge number of elements this
  duration applies to

mkvmerge needs to build an internal index from the various header
atoms describing the file location, size, timestamp & duration of each
sample. For tracks where there's only a single entry with a duration
of "1 sample" (PCM audio, but also certain types of video codecs) it
would insert a number of entries in the internal index equal to the
"number of entries" indicated by that one entry in the "sample
duration" table. This is fine for codec types with a low number of
samples per second like video with e.g. 25 frames per second.

However, with PCM audio and sample rates of up to 192.000 Hz this
would mean 192.000 entries per second in the internal index. At a size
of about 40 bytes per index entry an hour long PCM track would cause
mkvmerge to allocate an index of size 192.000 samples/s * 40
bytes/sample * 60 m/h * 60 s/m = 27.648.000.000 bytes = 25.75 GB of
RAM.

Temporarily it would even allocate twice as much while processing the
edit list atoms.

The fix is not to do this expansion for PCM audio at all. Instead
mkvmerge will now use the initially mentioned chunks of consecutive
samples for the internal index, which is also done for most other
types of audio & video codecs. For PCM this might mean a much lower
number of maybe hundreds to thousands, but not billions.

A side effect of the fix is several orders of magnitude fewer memory
allocations & file reading operations are done, speeding up PCM-in-MP4
processing quite a, too.

Part of the fix of #6149