Fix build error on macOS with FFmpeg 8
FFmpeg deprecated AV_CODEC_CAP_SUBFRAMES a while back.

This is fixed in Qt 6.9.2 which is a patch release
made on the top of Qt 6.9.1. As a patch release,
Qt 6.9.2 does not add any new functionality
but provides bug fixes and other improvements
and maintains both forward and backward compatibility
(source and binary) with the 6.9.x series.