mkvextract: AV1/IVF: fix type used in file header frame rate calculation
The fields in the file header are 32-bit wide, not 16-bit. However, as
the values are stored in Little Endian byte order (with least
significant bits first) the stored values are actually the calculated
16-bit values at the correct field addresses — it's just that the
values don't have to be clamped to 16 bits. In practice this should
have never made a difference; but it is better now.

Part of the fix of #6120.